docs: refactor the project documentation (#305)
### What

Refactor the project documentation file tree and update the texts. You can more easily check out the new docs structure at https://github.com/stellar/java-stellar-anchor-sdk/tree/docs-refactor.

Also, I created Readme.md files using symbolic links, so we can keep the numeric/sequential structure while still satisfying the Readme.md naming convension that github uses to identify the main markdown file in a folder.

### Why

The documentation was a bit disorganized and we want to have a clear path for partners willing to make use of it.

### Future Work

The next step is to add integration guidelines, so partners know how to develop their Anchor service and how to integrate it with the Anchor platform.