docs: add high-level database sizing information (#383)
### What

Update documentation with a high-level estimate of database storage usage for the Anchor Platform, considering the SEP-31 (with SEP-38) flow.

Our tests indicate that **each SEP-31 (with SEP-38) transaction occupies ~2KB of space in the database**. The tests also indicate that the data space usage tends to decrease as the number of SEP-31 transactions increases. These results were achieved by executing 500 SEP-31 transactions on a Postgres database.

### Why

Close #304.

### Known limitations

This is not a real-world scenario, but it's a good estimate of the amount of data that will be used by SEP-31 transactions.