Support SEP-31 non-happy-path where the Transaction is marked as `PENDING_CUSTOMER_INFO_UPDATE` (#448)
### What

Support SEP-31 non-happy-path where the Transaction is marked as `PENDING_CUSTOMER_INFO_UPDATE`. The changes include:
* Update the sequence diagram to better detail the error handling.
* Write integration tests for this use case.
* Write further unit tests for the PlatformAPI `TransactionService`.
* Improve the object schema conversion between database Sep31Transaction and Platform Api objects (also, including the appropriate tests)

### Why

We were not testing the case where the SEP-31 transaction failed to be completed by the Receiving Anchor backend and the Sending Anchor had to update the customer information. This flow was not clear in the Platform sequence diagram either.

Close #183.

### Known limitations

- [ ] #473
- [ ] #484