Feature: release checklist (#511)
### What

Add a release checklist to be followed before making a new release.

With this addition, we will have a new issue option called `Release a new Version!` when creating a [new issue](https://github.com/stellar/java-stellar-anchor-sdk/issues/new/choose). The idea is that we create a ticket for new releases and follow the checklist to make sure the release process is being followed thoroughly.

### Why

As part of the new gitflow process outlined in #487.