custody-server: SDFSOF-50 - Added additional validation to not start anchor platform if custody integration is enabled and account_creation or claimable_balances are enabled. Added additional tests (#914)
SDFSOF-50 - Added additional validation to not start anchor platform if custody integration is enabled and account_creation or claimable_balances are enabled. Added additional tests

What
Added additional validation to not start anchor platform if custody integration is enabled and account_creation or claimable_balances are enabled. Added additional tests

Why
As for now, integration with custody servers doesn't support automatic account creation. That's why, there was added an additional validation to not start anchor platform if custody integration is enabled and account_creation or claimable_balances are enabled