[ANCHOR-355] Implement SEP-6 deposit (#1035)
This implements the SEP-6 `GET /deposit` endpoint _mostly_ according to
the sequence diagram from the PR.

Where the implementation deviates is how `CUSTOMER_UPDATED` anchor
events are structured. This implementation sends out 1 event whenever a
SEP-12 customer is updated instead of fanning them out a
`TRANSACTION_STATUS_CHANGED` event for each ongoing transaction
submitted by the customer. This was done to keep the scope of this
change small, but we can consider whether to move this into the platform
later.