[ANCHOR-357] SEP-6: Implement deposit-exchange (#1125)
### Description

This implements the GET `deposit-exchange` endpoint. There's a lot of
duplicated code that could be shared with `deposit` and
`withdraw`/`withdraw-exchange`. Once the integration and end-to-end
tests have been implemented, I will focus on cleaning up the code.

### Context

SEP-6 implementation.

### Testing

`./gradlew test`

### Known limitations

Integration and end-to-end tests are missing, they will be added in
another PR.