[Anchor-418] Support auth with SEP10 in SEP38 (#1130)
<!-- If you're making a doc PR or something tiny where the below is
irrelevant, delete this
template and use a short description, but in your description aim to
include both what the
change is, and why it is being made, with enough context for anyone to
understand. -->

### Description

Add sep38.requires_sep10 configuration with default value to FALSE

### Context

Currently, according to the SEP-38 spec, some of the SEP38 endpoints
have SEP-10 as optional. This task seeks ensure we support this
optionality by add configurability to these endpoints' SEP-10
requirements.

### Testing
Tests were added to verify endpoints url pattern was added to filter if auth is
required