[ANCHOR-464] SEP-6: Enable client status callback (#1126)
### Description

This enables client status callback for SEP-6.

### Context

Status callback should be supported by the SEP-6 implement.

### Testing

`./gradlew test`

### Known limitations

The SEP-24 end-to-end was previously not asserting on the wallet status
callbacks. This was caused by the signature verification failing in the
wallet reference server, causing the GET `/callback` to always return
null. The test previously only asserted when the result was _not_ null.
These tests have been disabled in this PR due to flakiness but will be
re-enabled again when `develop` is merged in.