[ANCHOR-515] SEP-6: Remove KYC verification (#1177)
### Description

This removes KYC verification from the SEP-6 transaction initiation. Now
that KYC is removed, the customer may not be known to the Anchor at the
time of transaction initiation so the `customer` field cannot be set. To
allow Anchors to associate SEP transactions with a customer, the SEP-10
account memo is now added as a field to the `StellarId` object.

This means for the exchange endpoints, the platform can no longer make a
request to the Fee integration as it requires setting SEP-12 sender and
receiver fields. Business servers are expected to update the amounts
asynchronously like they already do for regular fees.

### Context

KYC verification is done "interactively" in SEP-6.

### Testing

- `./gradlew test`

### Documentation

Stellar docs need to be updated so that the `StellarId` object includes
an optional `memo` field

### Known limitations

N/A