[ANCHOR-388] SEP-6: Implement RPC actions (#1166)
### Description

This updates the RPC actions for SEP-6 and updates the reference server
implementation to use them. This PR is quite large as it updates the
unit tests for most of the RPC actions but, implementation-wise, not
much has changed. It might helpful to start by reviewing the
`Sep6End2EndTest` to get a sense of what changed with this PR.

### Context

SEP-6 should support RPC API.

### Testing

- `./gradlew test`

### Known limitations

- Platform API integration tests do not test the new flow but are
covered by the SEP-6 e2e tests.
- Refunds and custody service integration have not been implemented.

Both of these will be addressed by ANCHOR-508