[ANCHOR-587] Disable URL encoding of SEP-10 challenge response (#1266)
### Description

This commit fixes the URL encoding of the challenge transaction returned
by the SEP-10 endpoint. This is done by disabling the "HTML escaping"
done by the Gson instance used throughout the project. It is safe to do
so because Anchor Platform only returns HTTP responses with the
`application/json` content type, which does not require URL encoding.

### Context

The challenge transaction returned by the `/auth` endpoint is not
decodable.

### Testing

- `./gradlew test`

### Documentation

N/A

### Known limitations

N/A