[ANCHOR-484] Add 2.0 Helm Charts (#1294)
### Description

This adds Anchor Platform 2.x Helm charts. To support this change, it
required refactoring the Kafka and the reference server configurations.
The Helm Charts added by this PR should be considered a work in
progress. The values file is subject to change and needs to be
documented before being released publicly.

To test this, you should deploy all charts along with the Postgres, and
Kafka instances. You can modify the secret store to pass in credentials.

### Context

We are deploying the Anchor Platform to Kubernetes.

### Testing

- `./gradlew test`
- Testing with demo wallet.

### Documentation

This is missing READMEs and values file documentation. I will address
this in a follow-up PR to limit the LoC changes.

### Known limitations

N/A