[ANCHOR-698] Enable PaymentObservingAccountsManager for SEP-31 only (#1366)
### Description

This PR add check to `ObservingAccountsBeans ` to check if SEP-31 is
enabled. In this case the bean is always created, but only activate if
SEP-31 is enabled.

This is because this bean is required in the presence for the creation
of `PaymentObserverBeans`. We can't simply use conditional annotation
that disable the bean's creation

### Context

The PaymentObseringAccountManager should only be restarted when needed.
Currently only SEP-31 requires the service.

### Testing

- `./gradlew test`