[ANCHOR-712] Add transaction_id field to SEP-12 PUT/GET requests (#1383)
### Description

This adds the `transaction_id` fields to the SEP-12 GET/PUT endpoints.
This also updates the Customer callback implementation to return
different required fields if the transaction requires additional fields.
The required fields will be determined by a transaction "classifier"
service, which creates records in the KYC table whenever a transaction
requires additional KYC.

This implementation assumes that the KYC information is shared across
all customer transactions. This means that the `transaction_id` is
ignored by PUT /customer callback as KYC.

### Context

This is to support the new SEP-12 protocol and SEP-6 KYC flows. A
separate PR will update the flows and tests.

### Testing

- `./gradlew test`

### Documentation

Customer callback documentation will be updated with the
`transaction_id` field.

### Known limitations

N/A