[ANCHOR-477] Remove `withdraw_anchor_account` on SEP-24 txn (#1433)
### Description

This PR unsets the `withdraw_anchor_account` on SEP-24 withdraw
transaction creation. It is later when the transaction moves into
`pending_user_transfer_start`.

### Context

The deposit address can be something other than the asset distribution
account. It should not be set when the transaction is created.

### Testing

- `./gradlew test`

### Documentation

N/A

### Known limitations

N/A