[ANCHOR-764] Update NotifyCustomerInfoUpdated RPC (#1445)
### Description

Addresses
https://github.com/stellar/java-stellar-anchor-sdk/issues/1444.

This change:

- Updates the `NotifyCustomerInfoUpdated` RPC so that it updates the
transaction based on the requested Customer's status and sends a
Customer updated event. `customerId` and `type` are new parameters
needed to distinguish whether a customer is a SEP-31 receiver, sender,
or a SEP-6 customer.
- Updates SEP-12 service such that it passes the full SEP-12 customer
into the `AnchorEvent`. Previously only the customer ID was present in
the event. This is not a breaking change since the SEP-12 customer
object is serializer backward compatible since it also contains an ID.
- Sends a SEP-12 status callback to all clients.

Callback configuration and end-to-end tests will be added in another PR.

### Context

N/A

### Testing

- `./gradlew test`

### Documentation

stellar-docs PR will be created

### Known limitations

N/A