[ANCHOR-764] Add SEP-12 status callback tests (#1460)
### Description

This adds client callback URL configuration and updates the SEP-6
end-to-end tests to check for SEP-12 status updates. During the
implementation, I realized that the previously implemented
`RequestCustomerInfoUpdate` RPC is no longer needed after the recent
`NotifyCustomerInfoUpdate` change. I've marked the method as deprecated
and will remove it in 3.0.

Since this change added new routes to the Wallet server, I took this
opportunity to clean up the code a little.

### Context

This adds test coverage for a previous change.

### Testing

- `./gradlew test`

### Documentation

stellar-docs PR will be updated to reflect the
`NotifyCustomerInfoUpdate` field name change and
`RequestCustomerInfoUpdate` deprecation.

### Known limitations

N/A