[ANCHOR-796] Fix NotifyCustomerInfoUpdated RPC call (#1477)
### Description

This fixes the call to `NotifyCustomerInfoUpdated`. When the customer
does not exist prior to SEP-6 transaction creation, the reference server
needs to create one so that the ID is passed to the RPC. Otherwise, the
`NotifyCustomerInfoUpdated` to defaults setting the status to
`PENDING_ANCHOR`.

### Context

Demo wallet fix.

### Testing

- `./gradlew test`
- Tested deposit and withdrawal with demo wallet.

### Documentation

N/A

### Known limitations

We need to add tests that mirror the demo wallet flow.